Revert "Moves OldDoFn to runners-core"
This reverts commit 5f8b8c5b06cfd49c4293a20dff2eea99f1076444.
