+++ sdks/java/core/src/test/java/org/apache/beam/sdk/util/NameUtilsTest.java
@@ -42,18 +42,22 @@ public class NameUtilsTest {
 
   @Test
   public void testDropsStandardSuffixes() {
+    assertEquals("Embedded", NameUtils.approximateSimpleName("EmbeddedOldDoFn", true));
     assertEquals("Embedded", NameUtils.approximateSimpleName("EmbeddedDoFn", true));
     assertEquals("Embedded", NameUtils.approximateSimpleName("EmbeddedFn", true));
 
+    assertEquals("Embedded", NameUtils.approximateSimpleName("EmbeddedOldDoFn", false));
     assertEquals("Embedded", NameUtils.approximateSimpleName("EmbeddedDoFn", false));
     assertEquals("Embedded", NameUtils.approximateSimpleName("EmbeddedFn", false));
   }
 
   @Test
   public void testDropsStandardSuffixesInAllComponents() {
+    assertEquals("Embedded", NameUtils.approximateSimpleName("SomeOldDoFn$EmbeddedFn", true));
     assertEquals("Embedded", NameUtils.approximateSimpleName("SomeDoFn$EmbeddedDoFn", true));
     assertEquals("Embedded", NameUtils.approximateSimpleName("SomeFn$EmbeddedFn", true));
 
+    assertEquals("Some.Embedded", NameUtils.approximateSimpleName("SomeOldDoFn$EmbeddedFn", false));
     assertEquals("Some.Embedded", NameUtils.approximateSimpleName("SomeDoFn$EmbeddedDoFn", false));
     assertEquals("Some.Embedded", NameUtils.approximateSimpleName("SomeFn$EmbeddedFn", false));
   }
@@ -75,12 +79,12 @@ public class NameUtilsTest {
   /**
    * Inner class for simple name test.
    */
-  private class EmbeddedDoFn {
+  private class EmbeddedOldDoFn {
 
-    private class DeeperEmbeddedDoFn extends EmbeddedDoFn {}
+    private class DeeperEmbeddedOldDoFn extends EmbeddedOldDoFn {}
 
-    private EmbeddedDoFn getEmbedded() {
-      return new DeeperEmbeddedDoFn();
+    private EmbeddedOldDoFn getEmbedded() {
+      return new DeeperEmbeddedOldDoFn();
     }
   }
 
@@ -108,18 +112,18 @@ public class NameUtilsTest {
 
   @Test
   public void testSimpleName() {
-    assertEquals("Embedded", NameUtils.approximateSimpleName(new EmbeddedDoFn()));
+    assertEquals("Embedded", NameUtils.approximateSimpleName(new EmbeddedOldDoFn()));
   }
 
   @Test
   public void testAnonSimpleName() throws Exception {
-    assertEquals("Anonymous", NameUtils.approximateSimpleName(new EmbeddedDoFn() {}));
+    assertEquals("Anonymous", NameUtils.approximateSimpleName(new EmbeddedOldDoFn() {}));
   }
 
   @Test
   public void testNestedSimpleName() {
-    EmbeddedDoFn fn = new EmbeddedDoFn();
-    EmbeddedDoFn inner = fn.getEmbedded();
+    EmbeddedOldDoFn fn = new EmbeddedOldDoFn();
+    EmbeddedOldDoFn inner = fn.getEmbedded();
 
     assertEquals("DeeperEmbedded", NameUtils.approximateSimpleName(inner));
   }
